Modification of build scripts
