+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "translation_core",
   "productName": "translationCore",
-  "version": "0.1.2",
+  "version": "0.1.5",
   "description": "A bridge between TS and TM",
   "main": "main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "compile": "babel out --out-dir out",
-    "compile-modules": "babel modules --out-dir modules",
-    "build-win": "xcopy src out /Y /E && npm run compile && node scripts/build-win.js",
-    "build-mac": "cp -R src/ out/ && npm run compile && build --mac",
-    "build-lin": "cp src/* out/ && npm run compile && build --linux",
+    "compile": "babel src --out-dir src",
+    "build-win": "npm run compile && node scripts/build-win.js",
+    "build-winstaller": "node scripts/build-win-installer.js",
+    "build-mac": "npm run compile && build --mac",
+    "build-lin": "npm run compile && build --linux",
     "start": "electron ."
   },
   "build": {
@@ -36,6 +36,9 @@
   },
   "homepage": "https://github.com/unfoldingWord-dev/translationCore#readme",
   "devDependencies": {
+    "babel-core": "^6.13.2",
+    "babel-preset-react": "^6.11.0",
+    "babel-register": "^6.11.6",
     "electron": "^1.3.3",
     "electron-builder": "^5.26.0",
     "electron-packager": "^7.7.0",
@@ -46,9 +49,6 @@
   },
   "dependencies": {
     "babel-cli": "^6.11.4",
-    "babel-core": "^6.13.2",
-    "babel-preset-react": "^6.11.0",
-    "babel-register": "^6.11.6",
     "bootstrap": "^3.3.7",
     "event-emitter": "^0.3.4",
     "flux": "^2.1.1",
